requirements.txt
pytest==8.4.2
requests==2.32.5
selenium==4.35.0
python-dotenv
PyJWT
